basic ids integration with very basic microservices

+++ MicroservicesDemo/ProductsService/Migrations/20240305202527_InitialCreate.cs
@@ -21,7 +21,7 @@ namespace ProductsService.Migrations
                     Name = table.Column<string>(type: "nvarchar(250)", maxLength: 250, nullable: false),
                     Description = table.Column<string>(type: "nvarchar(2048)", maxLength: 2048, nullable: false),
                     Cost = table.Column<decimal>(type: "decimal(18,2)", nullable: false),
-                    StockQuantity = table.Column<double>(type: "float", nullable: false)
+                    StockQuantity = table.Column<long>(type: "bigint", nullable: false)
                 },
                 constraints: table =>
                 {
@@ -33,9 +33,9 @@ namespace ProductsService.Migrations
                 columns: new[] { "Id", "Cost", "Description", "Name", "StockQuantity" },
                 values: new object[,]
                 {
-                    { new Guid("0f5583e2-d5a3-491b-8e13-f57e04f46083"), 25m, "A pie from the past", "Perfectly Preserved Pie", 1.0 },
-                    { new Guid("7223634c-c992-4967-9cda-4cb4192f0a2e"), 10m, "For adding lube whenever you need", "Aluminum Oil Can", 100.0 },
-                    { new Guid("a0fa5a1f-fc38-4491-90da-2b04ea7bd679"), 5m, "Tasty beverage to kill your thirst", "Nuka-Cola", 10.0 }
+                    { new Guid("0f5583e2-d5a3-491b-8e13-f57e04f46083"), 25m, "A pie from the past", "Perfectly Preserved Pie", 1 },
+                    { new Guid("7223634c-c992-4967-9cda-4cb4192f0a2e"), 10m, "For adding lube whenever you need", "Aluminum Oil Can", 100 },
+                    { new Guid("a0fa5a1f-fc38-4491-90da-2b04ea7bd679"), 5m, "Tasty beverage to kill your thirst", "Nuka-Cola", 10 }
                 });
         }
